pyproject.toml
[tool.poetry]
name         = "agenticssdlc"
version      = "1.0.0"
description  = "Agentic SDLC — Production-grade multi-agent AI platform"
authors      = ["Samuele Cherubini"]
readme       = "README.md"
package-mode = false

[tool.poetry.dependencies]
python                  = "^3.11"
fastapi                 = "^0.115"
uvicorn                 = {extras = ["standard"], version = ">=0.25.0, <0.26.0"}
pydantic                = ">=2.7,<2.10"
pydantic-settings       = "^2.2"
langchain               = "^0.3"
langchain-core          = "^0.3"
langchain-google-vertexai = "^2.0"
langchain-chroma        = "^0.1"
langchain-qdrant        = "^0.1"
langgraph               = "^0.2"
chromadb                = "^0.5"
qdrant-client           = "^1.9"
sentence-transformers   = "^3.0"
torch                   = "^2.3"
google-cloud-storage    = "^2.16"
google-cloud-bigquery   = "^3.20"
supabase                = "^2.4"
prometheus-client       = "^0.20"
httpx                   = "^0.27"
python-snappy           = "^0.7"
chainlit                = "1.3.2"
bcrypt                  = "^4.1"

[tool.poetry.group.dev.dependencies]
pytest         = "^8.2"
pytest-asyncio = "^0.23"
pytest-cov     = "^5.0"
pre-commit     = "^3.7"
black          = "^24.4"
isort          = "^5.13"
flake8         = "^7.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.isort]
profile     = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
addopts      = "-v --tb=short"
pythonpath   = ["src"]

[tool.coverage.run]
omit = [
    "tests/*",
    "scripts/*",
    # Chainlit UI — requires live Chainlit server, covered by UI integration tests
    "src/ui/*",
    # Analytics ingest — requires live BigQuery/Supabase connections
    "src/analytics/*",
    # FastAPI app factory, middleware, routers — covered by integration tests
    "src/api/main.py",
    "src/api/middleware/*",
    "src/api/routers/*",
    # Grafana remote-write — requires live Prometheus/Grafana instance
    "src/core/core_remote_write.py",
    # RAG ingestion — requires live vector store (Chroma/Qdrant) connection
    "src/rag/rag_ingestion.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "def __repr__",
]
